VZ-7195: created status.go, moved some status-related functions
